requirements.txt
streamlit
python-dotenv
google.generativeai



